Refactor WithRole
