+++ packages/alfa-rules/src/sia-r55/rule.ts
@@ -1,4 +1,4 @@
-import { Diagnostic, Rule } from "@siteimprove/alfa-act";
+import { Rule } from "@siteimprove/alfa-act";
 import { DOM, Node, Role } from "@siteimprove/alfa-aria";
 import { Element, Namespace } from "@siteimprove/alfa-dom";
 import { Iterable } from "@siteimprove/alfa-iterable";
@@ -11,6 +11,7 @@ import * as dom from "@siteimprove/alfa-dom";
 import { expectation } from "../common/act/expectation";
 import { Group } from "../common/act/group";
 import { Question } from "../common/act/question";
+import { WithRole } from "../common/diagnostic/with-role";
 
 import { normalize } from "../common/normalize";
 
@@ -101,60 +102,4 @@ export namespace Outcomes {
         role
       )
     );
-}
-
-/**
- * @internal
- */
-export class WithRole extends Diagnostic {
-  public static of(message: string, role: Role.Name = "none"): WithRole {
-    return new WithRole(message, role);
-  }
-
-  private readonly _role: Role.Name;
-
-  private constructor(message: string, role: Role.Name) {
-    super(message);
-    this._role = role;
-  }
-
-  public get role(): Role.Name {
-    return this._role;
-  }
-
-  public equals(value: WithRole): boolean;
-
-  public equals(value: unknown): value is this;
-
-  public equals(value: unknown): boolean {
-    return (
-      value instanceof WithRole &&
-      value._message === this._message &&
-      value._role === this._role
-    );
-  }
-
-  public toJSON(): WithRole.JSON {
-    return {
-      ...super.toJSON(),
-      role: this._role,
-    };
-  }
-}
-
-/**
- * @internal
- */
-export namespace WithRole {
-  export interface JSON extends Diagnostic.JSON {
-    role: string;
-  }
-
-  export function isWithRole(value: Diagnostic): value is WithRole;
-
-  export function isWithRole(value: unknown): value is WithRole;
-
-  export function isWithRole(value: unknown): value is WithRole {
-    return value instanceof WithRole;
-  }
 }
